refactor(terminal): extract types to files
Fields on Frame that were private are now pub(crate).
